Flytte uttaksdatoer inn i uttaksinfo
