+++ src/uttaksplan/utils/stønadskontoUtils.ts
@@ -7,7 +7,7 @@ import {
 import { getAntallUkerFellesperiode } from 'uttaksplan/utils/permisjonUtils';
 import { Kjønn } from 'app/types/common';
 import { erFarEllerMedmor } from 'app/util/personUtil';
-import { SøkerGrunnlag } from 'uttaksplan/types/uttaksgrunnlag';
+import { SøkerGrunnlag } from 'uttaksplan/uttak/types';
 
 /** Forutsetter nå kun default som fødsel, ett barn og to foreldre */
 export const getTilgjengeligUttakEnkel = (
